Refactor HTTP-Redirect Sign method, Move test data to right folder
