+++ test/slo_logoutrequest_test.rb
@@ -1,5 +1,5 @@
 require File.expand_path(File.join(File.dirname(__FILE__), "test_helper"))
-require 'responses/logoutresponse_fixtures'
+require 'logout_responses/logoutresponse_fixtures'
 
 require 'onelogin/ruby-saml/slo_logoutrequest'
